+++ src/env-var/cc-env-var-editor-json.js
@@ -199,8 +199,12 @@ export class CcEnvVarEditorJson extends LitElement {
         .example code {
           background-color: var(--color-bg-neutral);
           border-radius: 0.25rem;
-          font-family: var(--cc-ff-monospace);
+          font-family: var(--cc-ff-monospace, monospace);
           padding: 0.15rem 0.3rem;
+        }
+
+        cc-input-text {
+          --cc-input-font-family: var(--cc-ff-monospace, monospace);
         }
       `,
     ];